pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image_api_openai"
version = "0.1.0"
authors = [
    {name = "Moha-Master", email = "hongkongreporter@outlook.com"},
]
description = "A proxy service converts ModelScope and SiliconFlow image APIs to OpenAI-compatible format"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "pydantic",
    "PyYAML",
]

[project.scripts]
image-api-openai = "image_api_openai.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["image_api_openai*"]

[tool.setuptools.package-data]
image_api_openai = ["config.yaml.example"]
